requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.1
cycler==0.12.1
deepface==0.0.93
distro==1.9.0
docx==0.2.4
et_xmlfile==2.0.0
filelock==3.13.1
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
fpdf==1.7.2
fsspec==2024.6.1
gast==0.6.0
gdown==5.2.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.1
groq==0.19.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jax==0.5.2
jaxlib==0.5.1
Jinja2==3.1.4
joblib==1.4.2
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.3.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
narwhals==1.30.0
networkx==3.3
numpy==2.2.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pyttsx3==2.98
pytz==2025.1
pywin32==309
referencing==0.36.2
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
scipy==1.15.2
sentencepiece==0.2.0
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SpeechRecognition==3.14.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
tf_keras==2.19.0
toml==0.10.2
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
whisper==1.1.10
wrapt==1.17.2
